+++ include/utils/customprint.hpp
@@ -10,12 +10,12 @@ using namespace std;
 
 // Functions to print variables
 void print(int x);
-void print(double x);
-void print(vector<double>);
-void print(vector<double>::size_type x);
+void print(long double x);
+void print(vector<long double>);
+void print(vector<long double>::size_type x);
 
 // Write to file
-void save_to_file(string file_path, vector<double>);
-void save_to_file(string file_path, vector< vector<double> >);
+void save_to_file(string file_path, vector<long double>);
+void save_to_file(string file_path, vector< vector<long double> >);
 
 #endif
