+++ include/utils/stats.hpp
@@ -11,23 +11,23 @@
 using namespace std;
 
 // Function to return maximum value of vector
-double vector_max(vector<double> v);
+long double vector_max(vector<long double> v);
 
 // Function to return minimum value of vector
-double vector_min(vector<double> v);
+long double vector_min(vector<long double> v);
 
 // Compute mean of a vector
-double mean(vector<double> x);
+long double mean(vector<long double> x);
 
 // Compute mean of a deque
-double stdev(deque<double> x);
+long double stdev(deque<long double> x);
 
 // Compute standard deviation of a list
-double stdev(deque<double> x);
+long double stdev(deque<long double> x);
 
 // Compute the x value for the optimium of a linear regression
-vector<double> quad_max(vector<double> x, vector<double> y);
+vector<long double> quad_max(vector<long double> x, vector<long double> y);
 
-vector<double> gauss(vector< vector<double> > A);
+vector<long double> gauss(vector< vector<long double> > A);
 
 #endif
