+++ src/agent_team/multipleteams.cpp
@@ -8,9 +8,9 @@ MultipleTeams::MultipleTeams(Parameters x){
     p = x;
 
     // Define the vectors that we need with the appropriate length
-    vector<double> temp1(static_cast <unsigned long> (p.max_iter/(double)p.n_agents), 0.0);
-    vector<vector <double> > temp2(static_cast <unsigned long> (p.n_reps), temp1);
-    vector<double> temp3(static_cast <unsigned long> (p.n_reps), 0.0);
+    vector<long double> temp1(static_cast <unsigned long> (p.max_iter/(long double)p.n_agents), 0.0);
+    vector<vector <long double> > temp2(static_cast <unsigned long> (p.n_reps), temp1);
+    vector<long double> temp3(static_cast <unsigned long> (p.n_reps), 0.0);
 
     best_solution = temp2;
     cdf = temp3;
@@ -22,16 +22,16 @@ MultipleTeams::MultipleTeams(string file_name){
     p.set_from_file(file_name);
 
     // Define the vectors that we need with the appropriate length
-    vector<double> temp1(static_cast <unsigned long> (p.max_iter/(double)p.n_agents), 0.0);
-    vector< vector<double> > temp2(static_cast <unsigned long> (p.n_reps), temp1);
-    vector<double> temp3(static_cast <unsigned long> (p.n_reps), 0.0);
+    vector<long double> temp1(static_cast <unsigned long> (p.max_iter/(long double)p.n_agents), 0.0);
+    vector< vector<long double> > temp2(static_cast <unsigned long> (p.n_reps), temp1);
+    vector<long double> temp3(static_cast <unsigned long> (p.n_reps), 0.0);
 
     best_solution = temp2;
     cdf = temp3;
 }
 
 //// This actually solves the problem LOTS of times.
-double MultipleTeams::solve(void){
+long double MultipleTeams::solve(void){
 
     for(int i=0; i<p.n_reps; i++) {
         // Instantiate a new team
