+++ src/utils/customprint.cpp
@@ -5,18 +5,18 @@ void print(int x){
     cout<<x<<endl;
 }
 
-//// Function to print a double
-void print(double x){
+//// Function to print a long double
+void print(long double x){
     cout<<x<<endl;
 }
 
 //// Function to print the size of a vector
-void print(vector<double>::size_type x){
+void print(vector<long double>::size_type x){
     cout<<x<<endl;
 }
 
 //// Function to print a vector. Black magic.
-void print(vector<double> x){
+void print(vector<long double> x){
     int end = static_cast <int> (x.size()-1);
     cout<<scientific<<"[ ";
     for(int i=0; i<x.size(); i++){
@@ -34,7 +34,7 @@ void print(vector<double> x){
 
 
 //// Save a vector to a file
-void save_to_file(string file_path, vector<double> x){
+void save_to_file(string file_path, vector<long double> x){
     // Open a file stream
     ofstream my_file;
 
@@ -55,7 +55,7 @@ void save_to_file(string file_path, vector<double> x){
 
 
 //// Save a vector of vectors to file. This happens a lot, actually.
-void save_to_file(string file_path, vector< vector<double> > x){
+void save_to_file(string file_path, vector< vector<long double> > x){
     // Open a file stream
     ofstream my_file;
